Have encoding, decoding fuctions in the same file.
